+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## [1.0.68] - 2026-02-24
+
+### Summary
+
+refactor(tests): deep code analysis engine with 5 supporting modules
+
+### Test
+
+- update test-results/.last-run.json
+
+### Other
+
+- update src/components/Chat.tsx
+- update src/core/bootstrap.ts
+- update src/plugins/system/logs.index.ts
+- update src/plugins/system/logsPlugin.test.ts
+- update src/plugins/system/logsPlugin.ts
+- update src/plugins/voice/index.ts
+- update src/plugins/voice/voiceCommandsPlugin.test.ts
+- update src/plugins/voice/voiceCommandsPlugin.ts
+
+
 ## [1.0.67] - 2026-02-24
 
 ### Summary